+++ actix-session/CHANGES.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Remove use of `async-trait` on `SessionStore` trait.
+- Minimum supported Rust version (MSRV) is now 1.75.
+
 ## 0.8.0
 
 - Set secure attribute when adding a session removal cookie.