+++ actix-web-httpauth/CHANGES.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Minimum supported Rust version (MSRV) is now 1.75.
+
 ## 0.8.1
 
 - Implement `From<Basic>` for `BasicAuth`.